move skew property extractor into transform utils
