+++ apps/builder/app/builder/features/style-panel/sections/transforms/transform-utils.test.ts
@@ -7,6 +7,7 @@ import {
   updateRotateOrSkewPropertyValue,
   updateTransformTuplePropertyValue,
   extractRotatePropertiesFromTransform,
+  extractSkewPropertiesFromTransform,
 } from "./transform-utils";
 import type { StyleInfo } from "../../shared/style-info";
 import {
@@ -310,6 +311,56 @@ describe("extractRotatePropertiesFromTransform", () => {
           ],
         },
         name: "rotateY",
+      },
+    });
+  });
+});
+
+describe("extractSkewPropertiesFromTransform", () => {
+  test("parses transform and returns undefined if no skew properties exists", () => {
+    expect(
+      extractSkewPropertiesFromTransform(
+        parseCssValue("transform", "rotateX(0deg) rotateY(0deg) scale(1.5)")
+      )
+    ).toEqual({ skewX: undefined, skewY: undefined });
+  });
+
+  test("parses transform and extracts valid skew properties", () => {
+    expect(
+      extractSkewPropertiesFromTransform(
+        parseCssValue(
+          "transform",
+          "skewX(10deg) skewY(20deg) rotate(30deg) scale(1.5)"
+        )
+      )
+    ).toEqual({
+      skewX: {
+        type: "function",
+        args: {
+          type: "layers",
+          value: [
+            {
+              type: "unit",
+              unit: "deg",
+              value: 10,
+            },
+          ],
+        },
+        name: "skewX",
+      },
+      skewY: {
+        type: "function",
+        args: {
+          type: "layers",
+          value: [
+            {
+              type: "unit",
+              unit: "deg",
+              value: 20,
+            },
+          ],
+        },
+        name: "skewY",
       },
     });
   });
