+++ packages/css-data/src/property-parsers/index.ts
@@ -1,4 +1,3 @@
 export * from "./transition";
 export * from "./transition-property-extractor";
 export * from "./shadow-properties-extractor";
-export * from "./skew-properties-extractor";
